+++ CHANGELOG.md
@@ -3,13 +3,26 @@
 The format is based on and uses the types of changes according to [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.3] - 2023-10-15
+## [1.0.3] - 2023-10-16
 
 ### Added
 
-- SPSWeather.ps1 - Add Installation process:
+- scripts/SPSWeather.ps1 - Add Installation process:
+
   - New parameters: Install, Uninstall and InstallAccount
   - New functions: Add-SPSSheduledTask and Remove-SPSSheduledTask
+
+- Wiki Documentation in repository - Add :
+  - wiki/Configuration.md
+  - wiki/Getting-Started.md
+  - wiki/Home.md
+  - wiki/Usage.md
+  - .github/workflows/wiki.yml
+
+### Changed
+
+- scripts/SPSWeather.ps1 - Remove ExclusionRules parameter
+- scripts/Config/CONTOSO-PROD.json - Add ExclusionRules parameter
 
 ## [1.0.2] - 2023-10-10