+++ RELEASE-NOTES.md
@@ -1,11 +1,24 @@
 # SPSWeather - Release Notes
 
-## [1.0.3] - 2023-10-15
+## [1.0.3] - 2023-10-16
 
 ### Added
 
-- SPSWeather.ps1 - Add Installation process:
+- scripts/SPSWeather.ps1 - Add Installation process:
+
   - New parameters: Install, Uninstall and InstallAccount
   - New functions: Add-SPSSheduledTask and Remove-SPSSheduledTask
+
+- Wiki Documentation in repository - Add :
+  - wiki/Configuration.md
+  - wiki/Getting-Started.md
+  - wiki/Home.md
+  - wiki/Usage.md
+  - .github/workflows/wiki.yml
+
+### Changed
+
+- scripts/SPSWeather.ps1 - Remove ExclusionRules parameter
+- scripts/Config/CONTOSO-PROD.json - Add ExclusionRules parameter
 
 A full list of changes in each version can be found in the [change log](CHANGELOG.md)